Common run method for writeTasks
We want to share a common "run" method for all write tasks since they
are all basically doing the same thing. To do this we have to change
write task to become generic and also include an implict
rowwriterfactory. In the future we may want to further extend this
so that metagenner integration will be easier.