Enhancements to the Read Side of the Tool
This patch adds in a mimic of some old PerfRow tests which were
implemented in the early days of DataStax. Although all the tests have
quite long names but we'll be able condense this at a later point. For
now this should be an entry point to testing the read performance of the
spark connector in a variety of circumstances.