Allow compiling against 4.7
Here we add in hooks so that we can add implicits based on version to
shade various tests. CassandraCount only exists after 1.2.4 because
prior to this it was the default behavior for all counts in 1.2.0. The
functionality doesn't exist at all in versions before that so we throw
an error if the function is attempted to be called.