Initial Streaming Test Commit
We have to do several changes to get Streaming to fit in correctly. The
biggest change is the addition fo the new StreamingTask and
StressReceiver. These pieces form the basis of what will hopefully be
many more future Streaming tests. The StressReceiver is a simple custom
receiver which spawns a Thread which uses a RowGenerator (as used in the
write tasks) to spawn C* rows. This allows us to use just about anything
that can generate Iterable[CassandraRow] style things to be Recevier
output.