Change Perfrow usage of Insant to j.u.Date
Older version of the SCC cannot handle a conversion from
java.util.Date (Driver native codec for timestamp columns) to
java.time.Insant. This means SCS doesn't work with older versions of DSE
or the SCC made before 3.0.

To fix this we switch to java.util.Date instead of instant